Learning code has been done more memory-efficient, in fact, the memory usage can be expected to be halved. The inference code still lacks required modifications. Stay tuned for the next commit. The code has been tested some, although it will be re-tested again in the next few days.

+++ 2D/CHOP/Prototype/vocabLearning/mergeIntoGraph.m
@@ -29,7 +29,7 @@ function [graph] = mergeIntoGraph(graph, level, ~, levelItr, position)
             for childItr = 1:numel(children)
                if ~ismember(newInstItr, previousLevel(children(childItr)).parents)
                     previousLevel(children(childItr)).parents = ...
-                        [previousLevel(children(childItr)).parents, newInstItr];
+                        [previousLevel(children(childItr)).parents, int32(newInstItr)];
                end
             end 
         end
